Fixed bad parsing in pushimm and related routines
extraction was ^push\s0x([a-f0-9]+)$, changed to ^push\s(0x[a-f0-9]+)$ to capture 0x chars, which were needed for some numerical comparisons on disp_to_dec routine. Had to also remove 0x again to prepare for zeropad after this. In the end zeropad shouldn't see the 0x if it is expecting hex input, but disp_to_dec works differently in this context. The above regex looked all the way to the final zeropad without considering the intermediate disp_to_dec.

Also checked related parsing (ignoring (not capturing) the 0x in the regex) in imm16imm8() and opimm8() to make sure these work as expected.